frozen mtp spec v2; drop v1 worker
