+++ python/sglang/test/kits/attention_unittest/runner_modes/speculative_draft_runner.py
@@ -29,7 +29,7 @@ from sglang.srt.speculative.frozen_kv_mtp_info import (
     FrozenKVMTPContext,
     FrozenKVMTPDraftInput,
 )
-from sglang.srt.speculative.frozen_kv_mtp_worker import FrozenKVMTPWorker
+from sglang.srt.speculative.frozen_kv_mtp_worker_v2 import FrozenDraftWorker
 from sglang.srt.speculative.spec_info import SpeculativeAlgorithm
 
 from ..attention_methods.dense_attention import (
@@ -226,26 +226,26 @@ class _FrozenKVMTPWorkerHarness:
         )
         self.model_runner.forward = model_forward
         self._hidden_size = settings.hidden_size
-        self.draft_forward = MethodType(FrozenKVMTPWorker.draft_forward, self)
+        self.draft_forward = MethodType(FrozenDraftWorker.draft_forward, self)
         self._frozen_kv_target_view = MethodType(
-            FrozenKVMTPWorker._frozen_kv_target_view,
+            FrozenDraftWorker._frozen_kv_target_view,
             self,
         )
         self._target_kv_pool_view = MethodType(
-            FrozenKVMTPWorker._target_kv_pool_view,
+            FrozenDraftWorker._target_kv_pool_view,
             self,
         )
-        self._set_positions = MethodType(FrozenKVMTPWorker._set_positions, self)
+        self._set_positions = MethodType(FrozenDraftWorker._set_positions, self)
         self._init_frozen_kv_metadata = MethodType(
-            FrozenKVMTPWorker._init_frozen_kv_metadata,
+            FrozenDraftWorker._init_frozen_kv_metadata,
             self,
         )
         self._init_frozen_kv_metadata_capture_cuda_graph = MethodType(
-            FrozenKVMTPWorker._init_frozen_kv_metadata_capture_cuda_graph,
+            FrozenDraftWorker._init_frozen_kv_metadata_capture_cuda_graph,
             self,
         )
         self._init_frozen_kv_metadata_replay_cuda_graph = MethodType(
-            FrozenKVMTPWorker._init_frozen_kv_metadata_replay_cuda_graph,
+            FrozenDraftWorker._init_frozen_kv_metadata_replay_cuda_graph,
             self,
         )
